Solutions to modify the CSS of MUI components
Based on the video in the comments.txt and the MUI documentation, we established around 9 solutions to overrides the CSS of MUI components. To sump up, here are the solutions:
- Solution 1: use props of components in the MUI API
- Solution 2: use the sx prop
- Solution 3: use the styled() function
- Solution 4: use the css function from @emotion/react
- Solution 5-6: use plain/global CSS (*.css) with custom classes and autogenerated MUI classes
- Solution 7-8: use module CSS (*.module.css) with custom classes and autogenerated MUI classes (which finally implies global)
- Solution 9: use a theme by modifying props of MUI components and palette for colors

We have to be carefull when we are modifying css of a component. We don't want to have side effects. The more precise we are in the selectors, the better the application will be.