Script generation
This script generates a file which allow us to have an overview of all countries as well as their sub-region and region in the world. More precisely, we can see:
- the name of the regions
- the name of the sub-regions for each region
- the name of the countries for each sub-region

NOTE: We create a second version of the countries all response. The number of data has not changed, only the display order. It's just to have a record of this.