WIP: Home + DetailsPage
We are implementing UI components to display the future data of the API. We are working with dummy data that has the correct architecture. When we are done implementing UI components, we will implement the redux logic to retrieve data from the API in order to wire it to our components.