WIP: DetailsPage
We planned to use the Grid component of MUI to display data of a country in the details page. However, the rowSpan property doesn't exists on this component. Because we planned to have some countries with multiple currencies and capitals and in order to have a beautiful UI, we need to use the Table MUI component.

We have seen that with the TableComponent we can have the same interface as with the GridComponent but we have the rowSpan property in bonus. Let's finish the configuration of this TableComponent instead of the GridComponent in order to display all data of our dummy country from the country.js file.

The dummy.js is just here to have an overview of countries that have:
- no/multiple native names (Every country has a native name but we tested it anyway).
- no/multiple currencies
- no/multiple capitals
- no/multiple languages
This file is temporary. Maybe we will put the code into a script file at some time.