DetailsPage done
The UI have been done for the details page. We know how to display each type of data. We've one a verification on some property to see if it exists but I think we should do it on every property to handle the case for each property. Let's wire up our home page and the detail page with the API. Let's use @redux-toolkit to handle the global state of this application.