WIP: React Leaflet
We are using the leaflet package to display countries on a (interactive) map. React Leaflet is an overlay of leaflet and is suitable for integration into a React app.
In this commit, we manage to display the map on the "/interactive-map" route. We used the GeoJSON component to display country borders from geo data (countries.json from https://datahub.io/core/geo-countries). We added a dummy switch to see how we can change dynamically data of the GeoJSON. Finally, we have set up a button that redirects to the details page of the country on which the user clicked.

Next step is to build an interactive tree on the left of the map. This tree will display countries of each subregion/region of the world.

PS: The observations.txt file is used to established which countries of the API are not included in the GeoJSON data file (countries.json) and vice-versa.