Mui tree view + interactions
We create a dummy switch to change dynamically the zoom on the map.
After that, we built the tree view from countries stored in redux. Finally, we add some interactions between the tree view and the map: when we click on a country in the tree view, a zoom on the country is done and its borders are highlighted.

Next step is to do the same thing for regions and continents.