Displaying borders for regions and subregions
We display the borders of each region/sub-region by retrieving the corresponding countries in the GeoJSON file. We also defined view options (latlng, zoom) for each region/sub-region in constants.js to have a correct display and a great navigation on the map.

We also need to check that lat/lng and zoom are correct for each country. We already established a list of countries that have a lat/lng problem (https://gitlab.com/restcountries/restcountries/-/issues/222). If something's wrong, we can override existing data for the moment.